+++ lib/yass/cli/runner.rb
@@ -33,6 +33,19 @@ module Yass
         config.stderr.puts "#{e.class.name}: #{e.message}"
         return 1
       end
+
+      def self.watch(config, argv:)
+        config.stdout.puts "Watching for changes..."
+        watcher = Filewatcher.new([config.src_dir, config.layout_dir, config.template_dir].map(&:to_s))
+        yield watcher if block_given?
+        watcher.watch do |changes|
+          files = changes.map { |f, _| Pathname.new(f).relative_path_from(config.root).to_s }
+          config.stdout.puts "Building #{files.join ", "}"
+          config.clear_cache!
+          Yass::CLI::Runner.build(config, argv: argv)
+        end
+        return 0
+      end
     end
   end
 end